Endpoint is working in acc, so defaultHistoryData is transfered to testdata.
