+++ src/signals/my-incidents/__test__/defaultHistoryData.ts
@@ -1,0 +1,18 @@
+/* SPDX-License-Identifier: MPL-2.0 */
+/* Copyright (C) 2022 Gemeente Amsterdam */
+export const defaultHistoryData = [
+  {
+    when: '2022-11-16T12:00:00+00:00',
+    what: 'UPDATE_LOCATION',
+    action: 'Locatie gewijzigd naar:',
+    description: 'Locatie is gepinned op kaart',
+    _signal: '00000000-0000-0000-0000-000000000000',
+  },
+  {
+    when: '2022-12-16T12:00:00+00:00',
+    what: 'checking',
+    action: 'Does this show?',
+    description: 'Party on',
+    _signal: '00000000-0000-0000-0000-000000000001',
+  },
+]
