+++ src/signals/my-incidents/__test__/index.ts
@@ -1,1 +1,2 @@
 export { providerMock } from './provider-mock'
+export { defaultHistoryData } from './defaultHistoryData'
