+++ src/signals/my-incidents/components/History/History.tsx
@@ -25,29 +25,8 @@ export interface Props {
   incident?: MyIncident
 }
 
-export const defaultHistoryData = [
-  {
-    when: '2022-11-16T12:00:00+00:00',
-    what: 'UPDATE_LOCATION',
-    action: 'Locatie gewijzigd naar:',
-    description: 'Locatie is gepinned op kaart',
-    _signal: '00000000-0000-0000-0000-000000000000',
-  },
-  {
-    when: '2022-12-16T12:00:00+00:00',
-    what: 'checking',
-    action: 'Does this show?',
-    description: 'Party on',
-    _signal: '00000000-0000-0000-0000-000000000001',
-  },
-]
-
 export const History = ({ incident }: Props) => {
-  const {
-    get,
-    data = defaultHistoryData,
-    error,
-  } = useFetch<HistoryInstance[]>()
+  const { get, data, error } = useFetch<HistoryInstance[]>()
   const location = useLocationReferrer() as Location
   const history = useHistory()
